backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-CORS==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
WTForms==3.0.1
cryptography==41.0.4
bcrypt==4.0.1
python-socketio==5.8.0
flask-socketio==5.3.6
eventlet==0.33.3
